refactor: Move PageNotFound component into MainPage components folder.
